requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-dotenv==1.0.0
pandas==2.0.3
PyYAML==6.0.1
lxml==4.9.3
nltk==3.8.1
textblob==0.17.1
